+++ UnityProject/Assets/Scripts/Detective/AppliedDetails.cs.meta
