+++ UnityProject/Assets/Scripts/Items/Weapons/Projectiles/Behaviours/ProjectileDecal.cs
@@ -1,4 +1,5 @@
-﻿using Items;
+﻿using Detective;
+using Items;
 using Mirror;
 using UnityEngine;
 
@@ -20,31 +21,28 @@ namespace Weapons.Projectiles.Behaviours
 			}
 
 			RpcClientSpawn(hit.HitWorld);
-
+			AppliedDetails AppliedDetails = null;
 			if (hit.CollisionHit.GameObject == null)
 			{
 				var Matrix =  MatrixManager.AtPoint(hit.TileHitWorld, isServer);
 				if (Matrix != null)
 				{
 					var Node = Matrix.MetaDataLayer.Get(hit.TileHitWorld.ToLocal(Matrix.Matrix).RoundToInt());
-					Node.AppliedDetail.AddDetail(new Detail()
-					{
-						CausedByInstanceID = 0,
-						Description = $"A bullet hole that looks like it was made by a {this.gameObject.name}",
-						DetailType = DetailType.BulletHole
-					});
+					AppliedDetails = Node.AppliedDetails;
 				}
 			}
 			else
 			{
 				var Node = hit.CollisionHit.GameObject.GetComponent<Attributes>();
-				Node.AppliedDetail.AddDetail(new Detail()
-				{
-					CausedByInstanceID = 0,
-					Description = $"A bullet hole that looks like it was made by a {this.gameObject.name}",
-					DetailType = DetailType.BulletHole
-				});
+				AppliedDetails = Node.AppliedDetails;
 			}
+
+			AppliedDetails.AddDetail(new Detail()
+			{
+				CausedByInstanceID = 0,
+				Description = $"A bullet hole that looks like it was made by a {this.gameObject.name}",
+				DetailType = DetailType.BulletHole
+			});
 
 			return false;
 		}
